invalid-imports/pyproject.toml
[tool.poetry]
name = "test-fleet-tg-import"
version = "0.1.0"
description = ""
authors = ["Vitali Falileev <vitali.falileev@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
python-telegram-bot = "^21.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
